Revert patch to increment mcycle every instruction.
It was buggy as WFI sets mcycle to mtimecmp which was broken since
mtime and mcycle no longer increment at the same rate.  It turns out
we don't actually need it anyway as we can use minstret to get the
number of retired instructions which is what we actually want.