Fix bad pattern match in isCapSentry.
Sail does not permit pattern matching on symbols except for constructors: instead it treats them as fresh variables matching a wildcard expression.
That means this match was totally broken and always returned true.
The only use of this function is in CJALR and the bug meant it was possible to jump to a capability that is not sealed with a sentry type.
We fix this by using guards as suggested by @bacam.
Fixes #5 .