Add stack high water mark.
This adds two new rw CSRs:
  - MSHWM   machine mode stack high water mark
  - MSHWMB  base for stack high water

For every store the physical address is checked to see if is in the
range MSHWMB (inclusive) to MSHWM (exclusive). If it is then MSHWM set
to the store address. This allows software to track the lowest stored
address on the stack which can be used to optimise stack clearing.