Make store-local violations non-trapping.
It's difficult to write defensive code if stores may trap depending on
the copied value.  Allowing store-local to trap makes it easy to force
anything that copies on behalf of a caller able to trap by inserting a
local capability into the source that it's copying from.

This change is already made in CHERIoT Ibex, the spec is catching up.

This also removes the associated exception codes.  We may wish to
reserve them for a debugging mode for finding places where the tag
clearing occurs.

Fixes #20.