Remove bounds checks on jumps / branch. Simplify and clarify MEPCC and MTCC behaviour.
As per #23 we want to delay the bounds check on jumps / branches until instruction
fetch in order to simplify hardware. Due to this potentially leading to
unrepresentable MEPCC values we also clear the tag of MEPCC on instruction fetch
bounds violations. Due to this this got a bit mixed up with #30 which clarifies
and simplifies validation and legalization of MEPCC / MTCC.