Fix reversed T and B fields in the capability encoding diagram.
There was an inconsistency between the Sail implementation and this
document about the locations of the T and B fields in the capability
encoding.  The document had the T and B fields swapped compared to the
Sail (which matches the Ibex implementation) so we treat the Sail as
canonical and update the document to match i.e. B is in bits 0 to 8 of
the metadata word and T is in 9 to 17.

Also fix a minor typo in description of R bit.