Require store local when storing return sentries (#64)
* Belatedly fix StoreCapImmediate docs

We no longer throw exceptions for !SL authorities storing !G capabilities, just
clear the tag.

* Require store local when storing return sentries

Backwards control-flow arcs should ideally be confined to the stack and register
save areas.  Conveniently, we have mechanism in the RTOS to identify exactly
those areas of memory, with capabilities bearing `SL` (store local) permission.
And, after https://github.com/microsoft/cheriot-sail/pull/54, the ISA has
mechanism for identifying backwards control-flow arcs, with the two return
sentry types.  We should have capability store impose the requirement for `SL`
in the authorizing cap if the cap being stored is a return sentry.

Credit where it's due: this is Robert's idea, originally suggested in the
obviously-wrong-in-retrospect
https://github.com/microsoft/cheriot-sail/issues/63 ("Have CJALR create !G
sentries?").

Co-authored-by: Robert Norton <robert.norton@microsoft.com>

* Review feedback

Co-authored-by: Robert Norton <1412774+rmn30@users.noreply.github.com>

---------

Co-authored-by: Robert Norton <robert.norton@microsoft.com>
Co-authored-by: Robert Norton <1412774+rmn30@users.noreply.github.com>