No functional change: simplify condition for jalr sentry types. (#117)
* No functional change: simplify condition for jalr sentry types.

I think this is easier to understand as the 'else' hides the tricky
cases and we keep the important ones (function call and return).  I
used Sail SMT to check that this is equivalent and also pulled it out
to a let variable for readability.

We could optionally give the table in the text the same treatment.