Default to using -smt_auto for property checking.
At the time this Makefile was written -smt_auto was broken but it now
works well and shows you the decoded counterexample which is nice.
Keep the old way around just in case.  Also allow choosing the solver
and default to z3 which appears to be much faster for our properties.